pyproject.toml
[project]
name = "doc-extractor"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "numpy>=2.4.3",
    "openai>=2.29.0",
    "pandas>=2.2,<3",
    "pymupdf>=1.27.2.2",
    "python-dotenv>=1.2.2",
    "scikit-learn>=1.8.0",
    "scipy>=1.17.1",
    "streamlit>=1.55.0",
]
